backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18
python-dotenv==1.0.1
openai>=1.58.0
qdrant-client==1.12.1
pypdf==5.1.0
langdetect==1.0.9
pydantic==2.9.2
